+++ Sources/MediaBridge/include/MediaBridge.h
@@ -1,3 +1,0 @@
-#ifndef MEDIA_BRIDGE_H
-#define MEDIA_BRIDGE_H
-#endif